+++ publishing-sdk/src/androidTest/java/com/ably/tracking/publisher/MapboxTests.kt
@@ -52,6 +52,7 @@ class MapboxTests {
             12345,
             true,
             null,
+            null,
         )
 
     private fun getLocationData(context: Context): LocationHistoryData {