+++ subscribing-sdk/src/main/java/com/ably/tracking/subscriber/CoreSubscriber.kt
@@ -32,6 +32,7 @@ internal interface CoreSubscriber {
     val rawLocations: SharedFlow<LocationUpdate>
     val trackableStates: StateFlow<TrackableState>
     val resolutions: SharedFlow<Resolution>
+    val nextLocationUpdateIntervals: SharedFlow<Long>
 }
 
 internal fun createCoreSubscriber(
@@ -59,6 +60,7 @@ private class DefaultCoreSubscriber(
     private val _enhancedLocations: MutableSharedFlow<LocationUpdate> = MutableSharedFlow(replay = 1)
     private val _rawLocations: MutableSharedFlow<LocationUpdate> = MutableSharedFlow(replay = 1)
     private val _resolutions: MutableSharedFlow<Resolution> = MutableSharedFlow(replay = 1)
+    private val _nextLocationUpdateIntervals: MutableSharedFlow<Long> = MutableSharedFlow(replay = 1)
 
     override val enhancedLocations: SharedFlow<LocationUpdate>
         get() = _enhancedLocations.asSharedFlow()
@@ -71,6 +73,9 @@ private class DefaultCoreSubscriber(
 
     override val resolutions: SharedFlow<Resolution>
         get() = _resolutions.asSharedFlow()
+
+    override val nextLocationUpdateIntervals: SharedFlow<Long>
+        get() = _nextLocationUpdateIntervals.asSharedFlow()
 
     init {
         val channel = Channel<Event>()
@@ -141,7 +146,6 @@ private class DefaultCoreSubscriber(
                         subscribeForChannelState()
                         subscribeForEnhancedEvents()
                         subscribeForRawEvents()
-                        subscribeForResolutionEvents()
                         event.callbackFunction(Result.success(Unit))
                     }
                     is PresenceMessageEvent -> {
@@ -150,6 +154,7 @@ private class DefaultCoreSubscriber(
                                 if (event.presenceMessage.data.type == ClientTypes.PUBLISHER) {
                                     properties.isPublisherOnline = true
                                     updateTrackableState(properties)
+                                    updatePublisherResolutionInformation(event.presenceMessage.data)
                                 }
                             }
                             PresenceAction.LEAVE_OR_ABSENT -> {
@@ -158,7 +163,11 @@ private class DefaultCoreSubscriber(
                                     updateTrackableState(properties)
                                 }
                             }
-                            else -> Unit
+                            PresenceAction.UPDATE -> {
+                                if (event.presenceMessage.data.type == ClientTypes.PUBLISHER) {
+                                    updatePublisherResolutionInformation(event.presenceMessage.data)
+                                }
+                            }
                         }
                     }
                     is ChangeResolutionEvent -> {
@@ -226,9 +235,10 @@ private class DefaultCoreSubscriber(
         }
     }
 
-    private fun subscribeForResolutionEvents() {
-        ably.subscribeForResolutionEvents(trackableId) {
-            scope.launch { _resolutions.emit(it) }
+    private fun updatePublisherResolutionInformation(presenceData: PresenceData) {
+        presenceData.resolution?.let { publisherResolution ->
+            scope.launch { _resolutions.emit(publisherResolution) }
+            scope.launch { _nextLocationUpdateIntervals.emit(publisherResolution.desiredInterval) }
         }
     }